requirements.txt
streamlit_authenticator
streamlit
python-dotenv
transformers>=4.41.2
torch
tenacity
pymupdf
pdfminer.six
crew
crewai
langchain==0.1.20
langchain-core
langchain-community
langchain-groq
tokenizers==0.19.0
threadpoolctl
scikit-learn
aiofiles<24.0.0,>=23.1.0
asyncer<0.0.3,>=0.0.2
fastapi-socketio<0.0.11,>=0.0.10
filetype<2.0.0,>=1.2.0
lazify<0.5.0,>=0.4.0
literalai
nest-asyncio<2.0.0,>=1.5.6
numpy
pyjwt<3.0.0,>=2.8.0
python-multipart<0.0.10,>=0.0.9
syncer<3.0.0,>=2.0.3
tomli<3.0.0,>=2.0.1
uptrace<2.0.0,>=1.22.0
uvicorn<0.26.0,>=0.25.0
watchfiles<0.21.0,>=0.20.0
dirtyjson<2.0.0,>=1.0.8
chardet
lxml
PyPDF2
pandas
huggingface_hub
datasets
pyYAML  
onnxruntime
fuzzywuzzy 
python-Levenshtein 
python-docx
spacy
plotly